[Test] MemberAdaptorImplTest 추가
